+++ packages/gill/src/core/index.ts
@@ -8,7 +8,7 @@ export * from "./send-and-confirm-transaction-with-signers";
 export * from "./prepare-transaction";
 export * from "./create-solana-client";
 export * from "./accounts";
-export * from "./keypairs";
+export * from "./keypairs-extractable";
 export * from "./base64-to-transaction";
 export * from "./base64-from-transaction";
 export * from "./simulate-transaction";
